save the last line of dom in a buffer variable to fix choppy cursor navigation
when the cursor is moving in the dom buffer, even on horizontal movements the function to
highlight the corresponding hex file region is called every time. this makes it laggy, as that
function is quite intensive. The patch caches the most recently highlighted line in the file buffer
and calls the function only if it has changed